fix ration screen bug with OSG
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@5041 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: ae193199463543dac00594b60ca82ea1767b3411
Former-commit-id: 832ff44f5f80395427c06ae6d59c340ee8ad6f43

+++ src/modules/graphic/osggraph/OsgMain.cpp
@@ -462,7 +462,7 @@ initView(int x, int y, int width, int height, int /* flag */, void *screen)
     m_sceneViewer->getCamera()->setName("Cam one");
     m_sceneViewer->getCamera()->setViewport(new osg::Viewport(0, 0, grWinw, grWinh));
     m_sceneViewer->getCamera()->setGraphicsContext(gw.get());
-    m_sceneViewer->getCamera()->setProjectionMatrixAsPerspective(67.5, ratio, 1, 12000.0);
+    m_sceneViewer->getCamera()->setProjectionMatrixAsPerspective(67.5, static_cast<double>(ratio), 1, 12000.0);
     m_sceneViewer->realize();
